requirements.txt
Django
djangorestframework
ifaddr
